# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';

export interface Activity {
  id: string;
  type: ActivityType;
  label: string;
  detail: string;
  createdAt: Date;
}