# lib/api-zod/src/generated/types/adDraft.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */
import type { AdDraftStatus } from './adDraftStatus';

export interface AdDraft {
  id: string;
  productId: string;
  productTitle: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  destinationUrl: string;
  status: AdDraftStatus;
  accountId: string;
  accountUsername: string;
  communityId: string;
  communityName: string;
  /** @nullable */
  publishedAt: Date | null;
  createdAt: Date;
}