# lib/api-zod/src/generated/types/adDraftInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */

export interface AdDraftInput {
  productId: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  accountId: string;
  communityId: string;
}