# lib/api-zod/src/generated/types/overview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { OverviewStoreStatus } from './overviewStoreStatus';

export interface Overview {
  storeStatus: OverviewStoreStatus;
  /** @nullable */
  storeName: string | null;
  productCount: number;
  draftCount: number;
  publishedCount: number;
  connectedAccountCount: number;
  publishedThisWeek: number;
  activity: Activity[];
}