# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IDDET Ads API for an embedded Shopify Dev App
 * OpenAPI spec version: 0.1.0
 */
import type { ProductAdStatus } from './productAdStatus';
import type { ProductStatus } from './productStatus';

export interface Product {
  id: string;
  title: string;
  description: string;
  price: number;
  currency: string;
  /** @nullable */
  imageUrl: string | null;
  status: ProductStatus;
  adStatus: ProductAdStatus;
  createdAt: Date;
}